Python/SchwarzLantern/requirements.txt
matplotlib==3.8.2
numpy==1.26.2
numpy-stl
